refactor components
Refactoring compoents by:
Making a components folder.
Added App.js to folder
Added CalcDisplay and
