+++ src/components/App.js
@@ -7,7 +7,9 @@
 
  */
 import React, { Component } from 'react';
-import './App.css';
+import '../App.css';
+import CalcButton from './CalcButton'
+import CalcDisplay from './CalcDisplay'
 
 class App extends Component {
   constructor(props) {
@@ -153,37 +155,6 @@ class App extends Component {
   }
 }
 
-class CalcDisplay extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-  render(){
-    return (
-      <React.Fragment>
-        <div className="index__displayCal" id="display">{this.props.disResult}</div>
-      </React.Fragment>
-    )
-  }
-}
 
-class CalcButton extends React.Component {
-  constructor(props) {
-    super(props);
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  handleChange() {
-    this.props.hChange(this.props.buttonDesc);
-  }
-
-  render() {
-      
-    return (
-      <React.Fragment>
-        <div id={this.props.buttonId} className="index__buttonBorder index__buttonClick" onClick={this.handleChange}>{this.props.buttonDesc}</div>
-      </React.Fragment>
-    )
-  }
-}
 
 export default App;
